+++ xprof/btx_aggreg_model.yaml
@@ -46,6 +46,11 @@
         :payload_field_class:
           :type: structure
           :members:
+            - :name: backend
+              :field_class:
+                :type: integer_signed
+                :field_value_range: 64
+                :cast_type: int64_t
             - :name: err_count
               :field_class:
                 :type: integer_signed
